requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiodns=3.0.0=pyhd8ed1ab_0
aiohttp=3.9.5=py312he37b823_0
aiosignal=1.3.1=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
bidict=0.23.1=pyhd8ed1ab_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.6.2=hf0a4a13_0
cffi=1.16.0=py312h8e38eb3_0
freetype=2.12.1=hadb7bae_2
frozenlist=1.4.1=py312he37b823_0
h11=0.14.0=pyhd8ed1ab_0
idna=3.7=pyhd8ed1ab_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcxx=17.0.6=h5f092b4_0
libdeflate=1.20=h93a5062_0
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libjpeg-turbo=3.0.0=hb547adb_1
libpng=1.6.43=h091b4b1_0
libsqlite=3.45.3=h091b4b1_0
libtiff=4.6.0=h07db509_3
libwebp-base=1.4.0=h93a5062_0
libxcb=1.15=hf346824_0
libzlib=1.3.1=hfb2fe0b_1
multidict=6.0.5=py312h670c8ac_0
ncurses=6.5=hb89a1cb_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.1=hfb2fe0b_0
pillow=10.3.0=py312h8a801b1_0
pip=24.0=pyhd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
pycares=4.0.0=py312h02f2b3b_3
pycparser=2.22=pyhd8ed1ab_0
python=3.12.3=h4a7b5fc_0_cpython
python-dotenv=1.0.1=pyhd8ed1ab_0
python-engineio=4.9.1=pyhd8ed1ab_0
python-socketio=5.11.2=pyhd8ed1ab_0
python_abi=3.12=4_cp312
readline=8.2=h92ec313_1
setuptools=70.0.0=pyhd8ed1ab_0
simple-websocket=1.0.0=pyhd8ed1ab_1
tk=8.6.13=h5083fa2_1
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
wheel=0.43.0=pyhd8ed1ab_1
wsproto=1.2.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yarl=1.9.4=py312he37b823_0
zstd=1.5.6=hb46c0d2_0
